[FEAT]: Create Layout For Doctor Profile and Detail Service
