+++ src/pages/MedicalService/ServiceDetail/components/Review/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Review';
