+++ src/pages/MedicalService/ServiceDetail/components/ReviewsSection/ReviewsSection.module.scss
@@ -1,0 +1,10 @@
+.cancelButton {
+    background-color: #6c757d;
+    border-color: #6c757d;
+    color: white;
+
+    &:hover {
+        background-color: #5a6268;
+        border-color: #545b62;
+    }
+}
