+++ src/pages/MedicalService/ServiceDetail/components/ReviewsSection/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ReviewsSection';
