+++ src/pages/MedicalService/ServiceDetail/components/ServiceProfileCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ServiceProfileCard';
